Fix selector refactor compile errors
Remove duplicate AddonStream import, add missing StreamSelectorSheet
import in DetailsScreen, and type updateStreamSelector session as Long
to match streamSelectorSession.

+++ android/app/src/main/java/com/crispy/tv/details/DetailsViewModel.kt
@@ -893,7 +893,7 @@ class DetailsViewModel internal constructor(
     }
 
     private fun updateStreamSelector(
-        session: Int,
+        session: Long,
         target: StreamLookupTarget,
         transform: (StreamSelectorUiState) -> StreamSelectorUiState,
     ) {
